+++ src/modules/equipment/domain/services/equipment.service.ts
@@ -11,8 +11,7 @@ export class EquipmentService {
   constructor(private readonly equipmentRepository: EquipmentRepository) {}
 
   async equipmentRegister(dto: EquipmentRegisterDTO) {
-    await this.equipmentRepository.Create(dto);
-    return true;
+    return await this.equipmentRepository.Create(dto);
   }
 
   async equipmentEdit(dto: EquipmentEditDTO) {
@@ -35,5 +34,32 @@ export class EquipmentService {
     await this.equipmentRepository.Delete(equipment.id);
 
     return true;
+  }
+
+  async findAll(page: number, limit: number) {
+    const skip = (page - 1) * limit;
+    const items = await this.equipmentRepository.findAll(skip, limit);
+
+    if (!items || items.length === 0) {
+      throw new NotFoundException({
+        statusCode: 404,
+        message: 'No equipment found',
+        error: 'Not found',
+      });
+    }
+    return items;
+  }
+
+  async getEquipmentById(id: string) {
+    const item = await this.equipmentRepository.findById(id);
+
+    if (!item) {
+      throw new NotFoundException({
+        statusCode: 404,
+        message: `Equipment with ID ${id} not found`,
+        error: 'Not found',
+      });
+    }
+    return item;
   }
 }
