+++ src/modules/equipment/domain/services/getlog.service.ts
